[MemorySSA] Be less aggressive with @llvm.lifetime.start
This is a workaround to keep LLVM from incorrectly removing stores in
rare cases. It has not yet been committed, and -- due to the direction
we seem to want to take MSSA in -- probably never will be.

It should be sufficient to work around the problem until we agree on a
good way forward.

Additional context: https://reviews.llvm.org/D43269

This is rebased to r327695 by gbiv@chromium.org.

Change-Id: I9cd22a13dddfb24613443f330005cad2264875fb
Signed-off-by: mydongistiny <jaysonedson@gmail.com>